+++ tests/Endpoint/EnvelopeTemplateDocumentsEndpointTest.php
@@ -20,6 +20,12 @@ class EnvelopeTemplateDocumentsEndpointTest extends EndpointTestCase
         self::assertLastRequest('GET', '/api/envelope-templates/bar/documents/foo/download?foo=bar');
     }
 
+    public function testPositions(): void
+    {
+        self::endpoint()->positions(['foo' => 'bar']);
+        self::assertLastRequest('PUT', '/api/envelope-templates/bar/documents/positions');
+    }
+
     protected static function endpoint(): EnvelopeTemplateDocumentsEndpoint
     {
         return self::dgs()->envelopeTemplates()->documents('bar');