+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## [Unreleased]
+### Added
+- Add `EnvelopeTemplateDocument.positions` endpoint
+
 ## [1.2.0] - 2021-08-06
 ### Added
 - Add `use` for `EnvelopeTemplateEndpoint`
@@ -79,4 +83,3 @@ see [UPGRADING-1.0](UPGRADING-1.0.md)
 - Add new IriTemplate and Iri classes to improve work with IRIs
 - Add more implementations of Iri, Add abstract ResourceIri to simplify Iri classes
 - Add more tests for increase code coverage
-